Cleanup null validation when configuration image
Also cleans up some tests and moves them around
