+++ jib-core/src/test/java/com/google/cloud/tools/jib/configuration/BuildConfigurationTest.java
@@ -29,11 +29,8 @@ import java.nio.file.Paths;
 import java.time.Instant;
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.HashMap;
-import java.util.Hashtable;
 import java.util.List;
 import java.util.Map;
-import java.util.TreeMap;
 import org.junit.Assert;
 import org.junit.Test;
 import org.mockito.Mockito;
@@ -214,91 +211,5 @@ public class BuildConfigurationTest {
           "base image configuration is required but not set and target image configuration is required but not set",
           ex.getMessage());
     }
-  }
-
-  @Test
-  @SuppressWarnings("JdkObsolete")
-  public void testBuilder_nullValues() {
-    // Java arguments element should not be null.
-    try {
-      BuildConfiguration.builder(Mockito.mock(JibLogger.class))
-          .setContainerConfiguration(
-              ContainerConfiguration.builder()
-                  .setProgramArguments(Arrays.asList("first", null))
-                  .build());
-      Assert.fail("The IllegalArgumentException should be thrown.");
-    } catch (IllegalArgumentException ex) {
-      Assert.assertNull(ex.getMessage());
-    }
-
-    // Entrypoint element should not be null.
-    try {
-      BuildConfiguration.builder(Mockito.mock(JibLogger.class))
-          .setContainerConfiguration(
-              ContainerConfiguration.builder().setEntrypoint(Arrays.asList("first", null)).build());
-      Assert.fail("The IllegalArgumentException should be thrown.");
-    } catch (IllegalArgumentException ex) {
-      Assert.assertNull(ex.getMessage());
-    }
-
-    // Exposed ports element should not be null.
-    try {
-      BuildConfiguration.builder(Mockito.mock(JibLogger.class))
-          .setContainerConfiguration(
-              ContainerConfiguration.builder()
-                  .setExposedPorts(Arrays.asList(new Port(1000, Protocol.TCP), null))
-                  .build());
-      Assert.fail("The IllegalArgumentException should be thrown.");
-    } catch (IllegalArgumentException ex) {
-      Assert.assertNull(ex.getMessage());
-    }
-
-    // Labels element should not be null.
-    Map<String, String> badLabels = new HashMap<>();
-    badLabels.put("label-key", null);
-    try {
-      ContainerConfiguration.builder().setLabels(badLabels);
-      Assert.fail("The IllegalArgumentException should be thrown.");
-    } catch (IllegalArgumentException ex) {
-      Assert.assertNull(ex.getMessage());
-    }
-
-    // Environment keys element should not be null.
-    Map<String, String> nullKeyMap = new HashMap<>();
-    nullKeyMap.put(null, "value");
-
-    try {
-      BuildConfiguration.builder(Mockito.mock(JibLogger.class))
-          .setContainerConfiguration(
-              ContainerConfiguration.builder().setEnvironment(nullKeyMap).build());
-      Assert.fail("The IllegalArgumentException should be thrown.");
-    } catch (IllegalArgumentException ex) {
-      Assert.assertNull(ex.getMessage());
-    }
-
-    // Environment values element should not be null.
-    Map<String, String> nullValueMap = new HashMap<>();
-    nullValueMap.put("key", null);
-    try {
-      BuildConfiguration.builder(Mockito.mock(JibLogger.class))
-          .setContainerConfiguration(
-              ContainerConfiguration.builder().setEnvironment(nullValueMap).build());
-      Assert.fail("The IllegalArgumentException should be thrown.");
-    } catch (IllegalArgumentException ex) {
-      Assert.assertNull(ex.getMessage());
-    }
-
-    // Can accept empty environment.
-    BuildConfiguration.builder(Mockito.mock(JibLogger.class))
-        .setContainerConfiguration(
-            ContainerConfiguration.builder().setEnvironment(ImmutableMap.of()).build());
-
-    // Environment map can accept TreeMap and Hashtable.
-    BuildConfiguration.builder(Mockito.mock(JibLogger.class))
-        .setContainerConfiguration(
-            ContainerConfiguration.builder().setEnvironment(new TreeMap<>()).build());
-    BuildConfiguration.builder(Mockito.mock(JibLogger.class))
-        .setContainerConfiguration(
-            ContainerConfiguration.builder().setEnvironment(new Hashtable<>()).build());
   }
 }
